Moved test cases to component/acceptance dir
